--- dashboard/src/components/DeploymentFormBody/BasicDeploymentForm/TabularSchemaEditorTable/IndeterminateCheckbox.tsx
@@ -1,20 +1,0 @@
-// Copyright 2022 the Kubeapps contributors.
-// SPDX-License-Identifier: Apache-2.0
-
-import { HTMLProps, useEffect, useRef } from "react";
-
-export default function IndeterminateCheckbox({
-  indeterminate,
-  className = "",
-  ...rest
-}: { indeterminate?: boolean } & HTMLProps<HTMLInputElement>) {
-  const ref = useRef<HTMLInputElement>(null!);
-
-  useEffect(() => {
-    if (typeof indeterminate === "boolean") {
-      ref.current.indeterminate = !rest.checked && indeterminate;
-    }
-  }, [ref, indeterminate, rest.checked]);
-
-  return <input type="checkbox" ref={ref} className={className + " cursor-pointer"} {...rest} />;
-}
